requirements.txt
openai>=1.3.0
google-generativeai>=0.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
qdrant-client>=1.16.0
requests>=2.31.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pdf2image>=1.17.0
python-multipart>=0.0.9
